python_ai/requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
pymongo==4.6.1
pyautogen==0.9.1a1
openai>=1.0.0
groq==0.4.1
flask-socketio>=5.3.2
eventlet>=0.33.0 